[Bronze V] Title: 이상한 기호, Time: 112 ms, Memory: 14484 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -16,7 +16,7 @@ public class Main {
         int N = Integer.parseInt(token.nextToken());
         int M = Integer.parseInt(token.nextToken());
 
-        int result = (N + M) * (N - M);
+        long result = (long) (N + M) * (N - M);
 
         bw.write(result + "");
